requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
PyMySQL==1.0.3
Werkzeug==2.2.3
email-validator==2.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
pymysql==1.0.3
cryptography==40.0.2